+++ 2024/opmodes/red/CenterstageRedLeftParking.java
@@ -24,7 +24,10 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
 TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
+package opmodes.red;
 
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.RobotLog;
@@ -40,35 +43,32 @@ import team25core.RobotEvent;
 import team25core.RunToEncoderValueTask;
 import team25core.SingleShotTimerTask;
 
-import com.acmerobotics.dashboard.FtcDashboard;
-import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
-import com.qualcomm.hardware.bosch.BNO055IMU;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.util.ElapsedTime;
-
-import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
-import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
-import org.opencv.core.*;
-import org.opencv.imgproc.Imgproc;
-import org.opencv.imgproc.Moments;
+import org.opencv.core.Mat;
+import org.opencv.core.MatOfPoint;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
 import org.openftc.easyopencv.OpenCvCameraRotation;
 import org.openftc.easyopencv.OpenCvPipeline;
 
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
+
+import org.opencv.core.*;
+import org.opencv.imgproc.Imgproc;
+import org.opencv.imgproc.Moments;
+
+
 import java.util.ArrayList;
 import java.util.List;
 
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 
-@Autonomous(name = "CenterstageBlueLeftDetect")
+@Config
+@Autonomous(name = "CenterstageRedLeftParking")
 //@Disabled
 
 //if any terms in the program are unknown to you, right click and press Go To > Declarations and Usages
-public class CenterstageBlueLeftDetect extends Robot {
+public class CenterstageRedLeftParking extends Robot {
 
 
     //wheels
@@ -81,55 +81,63 @@ public class CenterstageBlueLeftDetect extends Robot {
 
 
     //mechs
-    //private Servo servoMech;
-    private DcMotor motorMech;
-    private OneWheelDirectDrivetrain outtakeDrivetrain;
+//    private Servo servoMech;
     private DcMotor outtake;
+    private OneWheelDirectDrivetrain outtakeDrivetrain;
 
 
     //sensors
-    //private DistanceSensor distanceSensor;
-    //private DistanceSensorCriteria distanceSensorCriteria;
-    //private ColorSensor colorSensor;
+//    private DistanceSensor distanceSensor;
+//    private DistanceSensorCriteria distanceSensorCriteria;
+//    private ColorSensor colorSensor;
 
 
     //paths
-    private DeadReckonPath goToPark;
-    private DeadReckonPath goStraightToObject;
-    private DeadReckonPath goLeftToObject;
+    private DeadReckonPath goToParkFromMiddle;
+    private DeadReckonPath goToParkFromRight;
+    private DeadReckonPath goToParkFromLeft;
+
+    private DeadReckonPath goMiddleToObject;
     private DeadReckonPath goRightToObject;
+    private DeadReckonPath goLeftToObject;
+
     private DeadReckonPath outtakePath;
 
 
     //variables for constants
     //these constants CANNOT be changed unless edited in this declaration and initialization
-    static final double FORWARD_DISTANCE = 0;
-    static final double RIGHT_DISTANCE = 0;
-    static final double LEFT_DISTANCE = 28;
-    static final double DRIVE_SPEED = 0.6;
-    static final double OUTTAKE_DISTANCE = 1;
-    static final double OUTTAKE_SPEED = 0.6;
+    public static double FORWARD_DISTANCE = 14;
+    public static double RIGHT_DISTANCE = 14;
+    public static double LEFT_DISTANCE = 14;
+    public static double DRIVE_SPEED = 0.6;
+    public static double OUTTAKE_DISTANCE = 3;
+    public static double OUTTAKE_SPEED = 0.1;
 
 
     //telemetry
     private Telemetry.Item whereAmI;
-    private RunToEncoderValueTask motorMechTask;
-    private static final int DELAY = 5000;
+    private RunToEncoderValueTask outtakeTask;
     //integer 5000 represents 5000 milliseconds-change according to how long delay should be
+    private static final int DELAY = 5000;
 
-    //OPENCV DETECTION
+    public String objectDetectDirection;
 
-    double cX = 0;
-    double cY = 0;
-    double width = 0;
+    //private OpenCVRedDetectPipeline detectPipeline;
+
+    static double cX = 0;
+    static double cY = 0;
+    static double width = 0;
 
     private OpenCvCamera controlHubCam;  // Use OpenCvCamera class from FTC SDK
+
     private static final int CAMERA_WIDTH = 640; // width  of wanted camera resolution
     private static final int CAMERA_HEIGHT = 360; // height of wanted camera resolution
 
     // Calculate the distance using the formula
     public static final double objectWidthInRealWorldUnits = 3.75;  // Replace with the actual width of the object in real-world units
     public static final double focalLength = 728;  // Replace with the focal length of the camera in pixels
+
+
     /*
      * The default event handler for the robot.
      */
@@ -152,44 +160,59 @@ public class CenterstageBlueLeftDetect extends Robot {
     //initializes declared paths/tasks for the robot to do
     public void initPaths()
     {   //initializes the paths
-        goToPark = new DeadReckonPath();
-        goLeftToObject = new DeadReckonPath();
+        goToParkFromMiddle = new DeadReckonPath();
+        goToParkFromRight = new DeadReckonPath();
+        goToParkFromLeft = new DeadReckonPath();
+
+        goMiddleToObject = new DeadReckonPath();
         goRightToObject = new DeadReckonPath();
-        goStraightToObject = new DeadReckonPath();
+        goLeftToObject = new DeadReckonPath();
 
         //removes or clears the action of the paths
-        goToPark.stop();
-        goStraightToObject.stop();
-        goLeftToObject.stop();
+        goToParkFromMiddle.stop();
+        goToParkFromRight.stop();
+        goToParkFromLeft.stop();
+
+        goMiddleToObject.stop();
         goRightToObject.stop();
+        goLeftToObject.stop();
 
 
         outtakePath = new DeadReckonPath();
-        outtakePath.addSegment(DeadReckonPath.SegmentType.STRAIGHT, OUTTAKE_SPEED, 0.75);
+        outtakePath.addSegment(DeadReckonPath.SegmentType.STRAIGHT, OUTTAKE_DISTANCE, OUTTAKE_SPEED);
 
         //addSegment adds a new segment or direction the robot moves into
 
-        //drive path goToPark
-        //drive path goToPark moves the robot forward and then strafes left
-        goToPark.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, LEFT_DISTANCE, DRIVE_SPEED);
-        outtakePath.addSegment(DeadReckonPath.SegmentType.STRAIGHT, FORWARD_DISTANCE, DRIVE_SPEED);
+        //robot moves to the object in the right
+        goRightToObject.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, 2, -DRIVE_SPEED);
+        goRightToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 13, DRIVE_SPEED);
+        goRightToObject.addSegment(DeadReckonPath.SegmentType.TURN, 43, DRIVE_SPEED);
+        goRightToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 1, -DRIVE_SPEED);
 
-        //drive path goStraightToObject
-        //drive path goStraightToObject moves the robot forward
-        //goStraightToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, FORWARD_DISTANCE, DRIVE_SPEED);
+        //robot moves to the object in the middle
+        goMiddleToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 9, DRIVE_SPEED);
 
-        //drive path goLeftToObject
-        //drive path goLeftToObject turns the robot left and moves forward
-        //goLeftToObject.addSegment(DeadReckonPath.SegmentType.TURN, FORWARD_DISTANCE, -DRIVE_SPEED);
-        //goLeftToObject.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, RIGHT_DISTANCE, DRIVE_SPEED);
+        //robot moves to the object in the left
+        goLeftToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 12, DRIVE_SPEED);
+        goLeftToObject.addSegment(DeadReckonPath.SegmentType.TURN, 43, -DRIVE_SPEED);
+        goRightToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 1, -DRIVE_SPEED);
 
-        //drive path goRightToObject
-        //drive path goRightToObject turns the robot right and moves forward
-        //goRightToObject.addSegment(DeadReckonPath.SegmentType.TURN, FORWARD_DISTANCE, DRIVE_SPEED);
-        //goRightToObject.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, RIGHT_DISTANCE, -DRIVE_SPEED);
 
+        //after robot places pixel in the middle position, drives to the parking spot in backstage
+        goToParkFromMiddle.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, 14, -DRIVE_SPEED);
+        goToParkFromMiddle.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 14, DRIVE_SPEED);
+        goToParkFromMiddle.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, 100, DRIVE_SPEED);
+
+        //after robot places pixel in the right position, drives to the parking spot in backstage
+        goToParkFromRight.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, LEFT_DISTANCE, -DRIVE_SPEED);
+        goToParkFromRight.addSegment(DeadReckonPath.SegmentType.STRAIGHT, 50, DRIVE_SPEED);
+
+        //after robot places pixel in the left position, drives to the parking spot in backstage
+        goToParkFromLeft.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, RIGHT_DISTANCE, DRIVE_SPEED);
+        goToParkFromLeft.addSegment(DeadReckonPath.SegmentType.TURN, 86, DRIVE_SPEED);
+        goToParkFromLeft.addSegment(DeadReckonPath.SegmentType.STRAIGHT, FORWARD_DISTANCE, DRIVE_SPEED);
         //initializes motorMechTask
-        //motorMechTask = new RunToEncoderValueTask(this, motorMech, 0, 0);
+//        outtakeTask = new RunToEncoderValueTask(this, outtake, 0, 0);
     }
 
     //initializes the declared motors and servos
@@ -204,7 +227,7 @@ public class CenterstageBlueLeftDetect extends Robot {
 
         //initializes the servo
         //servos are not in initPaths() because they do not get tasks unless a task is created for them in a specified method
-        //servoMech = hardwareMap.servo.get("servoMech");
+//        servoMech = hardwareMap.servo.get("servoMech");
 
 
         //sets wheel motors to run using the encoders
@@ -218,73 +241,125 @@ public class CenterstageBlueLeftDetect extends Robot {
         drivetrain.resetEncoders();
         drivetrain.encodersOn();
 
+        //initializes pipeline for openCV
+        //detectPipeline = new OpenCVRedDetectPipeline();
+
         //displays telemetry of robot location
         whereAmI = telemetry.addData("location in code", "init");
 
         //initializes motor mechanism, returns what motor would do if 0 power behavior was implemented on it,
         //rests encoder, and prepares motors to run on the encoders
-        //motorMech = hardwareMap.get(DcMotor.class, "motorMech");
-        //motorMech.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        //motorMech.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        //motorMech.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        outtake = hardwareMap.get(DcMotor.class, "outtake");
+        outtake.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        outtake.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        outtake.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        //initializes the outtake drivetrain, resets encoders, and prepares motor(s) to run on the encoders
+        //initializes the motor drivetrain, resets encoders, and prepares motor(s) to run on the encoders
         outtakeDrivetrain = new OneWheelDirectDrivetrain(outtake);
         outtakeDrivetrain.resetEncoders();
         outtakeDrivetrain.encodersOn();
 
 
         //initializes the color sensor and distance sensor for usage
-        //colorSensor = hardwareMap.get(RevColorSensorV3.class, "colorSensor");
-        //distanceSensor = hardwareMap.get(Rev2mDistanceSensor.class, "distanceSensor");
+//        colorSensor = hardwareMap.get(RevColorSensorV3.class, "colorSensor");
+//        distanceSensor = hardwareMap.get(Rev2mDistanceSensor.class, "distanceSensor");
 
-        //openCv
-        //initOpenCV();
+
+        initOpenCV();
         FtcDashboard dashboard = FtcDashboard.getInstance();
         telemetry = new MultipleTelemetry(telemetry, dashboard.getTelemetry());
         FtcDashboard.getInstance().startCameraStream(controlHubCam, 30);
+
+        telemetry.addData("Coordinate", "(" + (int) cX + ", " + (int) cY + ")");
+        telemetry.addData("Distance in Inch", (getDistance(width)));
+        telemetry.update();
 
         //calls method to start the initialization
         initPaths();
 
 
     }
-    //method that turns robot positioning depending on results from openCV
-    /*public void objectDetectionPath()
+
+    private void initOpenCV() {
+
+        // Create an instance of the camera
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier(
+                "cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+
+        // Use OpenCvCameraFactory class from FTC SDK to create camera instance
+        controlHubCam = OpenCvCameraFactory.getInstance().createWebcam(
+                hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
+
+        controlHubCam.setPipeline(new RedBlobDetectionPipeline());
+
+        controlHubCam.openCameraDevice();
+        controlHubCam.startStreaming(CAMERA_WIDTH, CAMERA_HEIGHT, OpenCvCameraRotation.UPRIGHT);
+    }
+
+    //moves to detected object and releases the pixel
+    public void moveToObjectAndReleasePixel(DeadReckonPath path)
     {
-        if (findPosition().equals("right")) {
-            //drive path goRightToObject
-            //drive path goRightToObject turns the robot right and moves forward
-            goRightToObject.addSegment(DeadReckonPath.SegmentType.TURN, FORWARD_DISTANCE, DRIVE_SPEED);
-            goRightToObject.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, RIGHT_DISTANCE, -DRIVE_SPEED);
-        }
-        else if (findPosition().equals("left")) {
-            //drive path goLeftToObject
-            //drive path goLeftToObject turns the robot left and moves forward
-            goLeftToObject.addSegment(DeadReckonPath.SegmentType.TURN, FORWARD_DISTANCE, -DRIVE_SPEED);
-            goLeftToObject.addSegment(DeadReckonPath.SegmentType.SIDEWAYS, RIGHT_DISTANCE, DRIVE_SPEED);
 
-        }
-        else {
-            //drive path goStraightToObject
-            //drive path goStraightToObject moves the robot forward
-            goStraightToObject.addSegment(DeadReckonPath.SegmentType.STRAIGHT, FORWARD_DISTANCE, DRIVE_SPEED);
-
-        }
-
-    }*/
-    //method that starts moving robot for path1 indicated above
-    public void goToPark()
-    {
-        this.addTask(new DeadReckonTask(this, goToPark, drivetrain ){
+        this.addTask(new DeadReckonTask(this, path, drivetrain ){
             @Override
             public void handleEvent(RobotEvent e) {
                 DeadReckonEvent path = (DeadReckonEvent) e;
                 if (path.kind == EventKind.PATH_DONE)
                 {
-                    RobotLog.i("Drove to the left");
-                    whereAmI.setValue("Parked on the left");
-                    delay(0);
+                    RobotLog.i("Drove to the object");
+                    whereAmI.setValue("At the object");
+                    releaseOuttake();
+                    if(findPosition().equals("center"))
+                    {
+                        goToPark(goToParkFromMiddle);
+                    }
+                    else if(findPosition().equals("right"))
+                    {
+                        goToPark(goToParkFromRight);
+                    }
+                    else
+                    {
+                        goToPark(goToParkFromLeft);
+                    }
+//                    delay(0);
+
+
+                }
+            }
+        });
+    }
+
+
+
+    public void detectObject()
+    {
+        if(findPosition().equals("right"))
+        {
+            moveToObjectAndReleasePixel(goRightToObject);
+        }
+        else if(findPosition().equals("center"))
+        {
+            moveToObjectAndReleasePixel(goMiddleToObject);
+        }
+        else
+        {
+            moveToObjectAndReleasePixel(goLeftToObject);
+        }
+    }
+
+    //robot goes to the backstage parking
+    public void goToPark(DeadReckonPath path)
+    {
+
+        this.addTask(new DeadReckonTask(this, path, drivetrain ){
+            @Override
+            public void handleEvent(RobotEvent e) {
+                DeadReckonEvent path = (DeadReckonEvent) e;
+                if (path.kind == EventKind.PATH_DONE)
+                {
+                    RobotLog.i("Drove to parking");
+                    whereAmI.setValue("In backstage parking");
+//                    delay(0);
 
 
                 }
@@ -309,67 +384,41 @@ public class CenterstageBlueLeftDetect extends Robot {
 
     //provides a certain task movement for the motor mech and displays telemetry stating robot is
     //executing the motor mech task
-    /*private void goMoveMotorMech() {
-        this.addTask(new DeadReckonTask(this, outtake, motorDrivetrain) {
+    private void releaseOuttake() {
+        this.addTask(new DeadReckonTask(this, outtakePath, outtakeDrivetrain) {
             @Override
             public void handleEvent(RobotEvent e) {
                 DeadReckonEvent path = (DeadReckonEvent) e;
                 if (path.kind == EventKind.PATH_DONE) {
-                    whereAmI.setValue("moved motor mech");
+                    whereAmI.setValue("released purple pixel");
 
                 }
             }
         });
-    }*/
+    }
 
     //provides certain movement for servo mechanism and displays telemetry stating robot
     //executed the servo task
-    /*private void setServoMech() {
-        servoMech.setPosition(0);
-        whereAmI.setValue("servo moved");
-    }*/
+//    private void setServoMech() {
+//        servoMech.setPosition(0);
+//        whereAmI.setValue("servo moved");
+//    }
 
 
 
-    //method that executes when the driver presses start on the driver station
+    //executes parking and releases pixel
     @Override
     public void start()
     {
         whereAmI.setValue("in Start");
-        goToPark();
-        //addTask(motorMechTask);
-
-        telemetry.addData("Coordinate", "(" + (int) cX + ", " + (int) cY + ")");
-        //telemetry.addData("Distance in Inch", (getDistance(width)));
-        //telemetry.addData("Position: ", findPosition());
-        telemetry.update();
-
+        detectObject();
 
 
     }
 
 
-
-
-    /////////////////////////////////// OPENCV STUFF ///////////////////////////////////////////////
-/*
-    private void initOpenCV() {
-
-        // Create an instance of the camera
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier(
-                "cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-
-        // Use OpenCvCameraFactory class from FTC SDK to create camera instance
-        controlHubCam = OpenCvCameraFactory.getInstance().createWebcam(
-                hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
-
-        controlHubCam.setPipeline(new RedBlobDetectionPipeline());
-
-        controlHubCam.openCameraDevice();
-        controlHubCam.startStreaming(CAMERA_WIDTH, CAMERA_HEIGHT, OpenCvCameraRotation.UPRIGHT);
-    }
-
-    public static class RedBlobDetectionPipeline extends OpenCvPipeline {
+    public static class RedBlobDetectionPipeline extends OpenCvPipeline
+    {
         @Override
         public Mat processFrame(Mat input) {
             // Preprocess the frame to detect red regions
@@ -458,7 +507,7 @@ public class CenterstageBlueLeftDetect extends Robot {
         double distance = (objectWidthInRealWorldUnits * focalLength) / width;
         return distance;
     }
-    private String findPosition(){
+    private static String findPosition(){
         if (cX > 400) {
             return "right";
         }
@@ -469,6 +518,13 @@ public class CenterstageBlueLeftDetect extends Robot {
             return "left";
         }
     }
-*/
+
+
 
 }
+
+
+//OpenCVSTuFF
+
+
+
