+++ 2024/opmodes/red/CenterstageRedRightParking.java
@@ -24,7 +24,7 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
 TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
 THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package opmodes;
+package opmodes.red;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.hardware.motors.RevRoboticsCoreHexMotor;
